analyzer.py
"""
Core module for scraping and analyzing pharmaceutical press releases.
"""
import os
import json
import time
import requests
import logging
from typing import List, Dict, Optional
from bs4 import BeautifulSoup
import pandas as pd
import google.generativeai as genai

logger = logging.getLogger(__name__)
logging.basicConfig(level=logging.INFO,
                    format="%(asctime)s - %(levelname)s - %(message)s")

class PharmaPressReleaseAnalyzer:
    """Main class for pharmaceutical press release analysis."""

    def __init__(self, serp_api_key: str, gemini_api_key: str):
        self.serp_api_key = serp_api_key
        self.gemini_api_key = gemini_api_key
        genai.configure(api_key=self.gemini_api_key)
        self.model = genai.GenerativeModel('gemini-2.0-flash')
        self.headers = {
            'User-Agent': (
                'Mozilla/5.0 (Windows NT 10.0; Win64; x64) '
                'AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
            ),
        }

    def search_pharmaceutical_press_release_urls(self) -> List[str]:
        """Return press-release URLs via SerpAPI or fallback."""
        # Implementation...
        return []

    # ... rest of methods refactored similarly, each with docstrings and error handling
