requirements[1].txt
beautifulsoup4>=4.12
google-generativeai>=0.1.0
google-search-results>=2.4.0
pandas>=1.5.0
requests>=2.28.0
